+++ projects/hipfft/clients/tests/hipfftw_test.cpp
@@ -1695,7 +1695,7 @@ namespace
                 = std::accumulate(batches.begin(),
                                   batches.end(),
                                   static_cast<ptrdiff_t>(1),
-                                  [](ptrdiff_t& acc, ptrdiff_t x) {
+                                  [](ptrdiff_t acc, ptrdiff_t x) {
                                       return acc * std::max(ptrdiff_t(1), std::abs(x));
                                   });
             const auto placement = get_random_element_in(place_range);
@@ -2082,7 +2082,7 @@ namespace
             {
                 GTEST_FAIL() << "undefined function pointers detected. Error info: " << e.what();
             }
-            catch(const std::runtime_error e)
+            catch(const std::runtime_error& e)
             {
                 if(log_content.empty() && exception_logger)
                     log_content = exception_logger->get_log();
@@ -2976,14 +2976,14 @@ namespace
             {
                 GTEST_FAIL() << "undefined function pointers detected. Error info: " << e.what();
             }
-            catch(const hip_runtime_error e)
+            catch(const hip_runtime_error& e)
             {
                 if(skip_runtime_fails)
                     GTEST_SKIP() << e.what() << "\nError code: " << e.hip_error << ".";
                 else
                     GTEST_FAIL() << e.what() << "\nError code: " << e.hip_error << ".";
             }
-            catch(const std::runtime_error e)
+            catch(const std::runtime_error& e)
             {
                 GTEST_FAIL() << e.what();
             }